fix: remove hash tests from file_utilities
These hash test fail depending on which OS they are run on, possibly because the newline ending is different. Windows uses \r\n, while Unix uses \n. This is a problem because the hash tests are comparing the hash of the file to a known hash. The hash is calculated based on the file contents, so if the newline ending is different, the hash will be different. This is why the hash tests fail on some OSs.